requirements.txt
apptools==4.5.0
autopep8==1.5.2
configobj==5.0.6
cycler==0.10.0
envisage==4.9.2
kiwisolver==1.2.0
matplotlib==3.2.1
mayavi==4.7.1
numpy==1.18.3
pycodestyle==2.5.0
pyface==6.1.2
Pygments==2.6.1
pyparsing==2.4.7
PySide2==5.14.2
python-dateutil==2.8.1
shiboken2==5.14.2
six==1.14.0
traits==6.0.0
traitsui==6.1.3
vtk==8.1.2
